a maximum of 4294967294 entities is enough for us -> use u32 instead of size_t for entity ids. this especially brings performance in x64 builds

+++ EntityTest/EntityTest.cpp
@@ -29,7 +29,7 @@ int main()
 
 		const auto start = std::chrono::high_resolution_clock::now();
 
-		for( size_t i = 0; i < ecs.MaxSize; i++ )
+		for( u32 i = 0; i < ecs.MaxSize; i++ )
 		{
 			auto entity = ecs.Create();
